Add user index for admin

+++ public/pages/UserIndex.jsx
@@ -1,6 +1,7 @@
 import { UserList } from "../cmps/UserList.jsx"
 import { userService } from "../services/user.service.js"
 import { showSuccessMsg, showErrorMsg } from '../services/event-bus.service.js'
+import { bugService } from "../services/bug.service.remote.js"
 
 
 const { useState, useEffect, useRef } = React
@@ -17,13 +18,25 @@ export function UserIndex() {
     }
 
     function onRemoveUser(userId) {
-        userService.remove(userId)
-            .then(() => {
-                const UsersToUpdate = users.filter(user => user._id !== userId)
-                setUsers(UsersToUpdate)
-                showSuccessMsg('user removed')
+        console.log(userId)
+        bugService.getUserBugs(userId)
+            .then(res => {
+                if (res.length > 0) {
+                    console.log('no no no')
+                    showErrorMsg(`User has bugs!!!`)
+                    throw new Error('no!')
+                }
+                userService.remove(userId)
+                    .then(() => {
+                        const UsersToUpdate = users.filter(user => user._id !== userId)
+                        setUsers(UsersToUpdate)
+                        showSuccessMsg('user removed')
+                    })
+                    .catch((err) => showErrorMsg(`Cannot remove user`, err))
             })
-            .catch((err) => showErrorMsg(`Cannot remove user`, err))
+        // .catch(err => showErrorMsg(`Cannot remove user`, err))
+
+
     }
 
     return (
